rustup: update to `nightly-2024-09-01` (~1.82).

+++ crates/rustc_codegen_spirv/src/builder/builder_methods.rs
@@ -56,7 +56,7 @@ macro_rules! simple_op {
                                     _ => return None,
                                 };
                                 Some(if signed {
-                                    size.sign_extend(x)
+                                    size.sign_extend(x) as u128
                                 } else {
                                     size.truncate(x)
                                 })
